feat(keypairs): add new type to replace FullId - This type is more in tune with the needs of a node. - NB: We will still need to evaluate how to best compose this.
