+++ src/identity/node.rs
@@ -10,7 +10,7 @@
 use crate::keys::{BlsKeypairShare, SignatureShare};
 use crate::{utils, Ed25519Digest, Error, PublicKey, Signature, XorName};
 use bls::{
-    serde_impl::SerdeSecret, PublicKeyShare as BlsPublicKeyShare,
+    serde_impl::SerdeSecret, PublicKeySet, PublicKeyShare as BlsPublicKeyShare,
     SecretKeyShare as BlsSecretKeyShare,
 };
 use ed25519_dalek::{Keypair as Ed25519Keypair, PublicKey as Ed25519PublicKey};
@@ -192,5 +192,118 @@ impl Debug for PublicId {
 impl Display for PublicId {
     fn fmt(&self, formatter: &mut Formatter) -> fmt::Result {
         Debug::fmt(self, formatter)
+    }
+}
+
+/// This is used at a network node for holding the
+/// obligatory Ed25519 keypair needed as Adult, and
+/// then a BLS keypair share when being promoted to Elder.
+/// (Also the corresponding public keys).
+/// The Ed25519 is kept as Elder, in case it is demoted.
+#[derive(Serialize, Deserialize)]
+pub struct NodeKeypairs {
+    ed25519: Ed25519Keypair,
+    bls: Option<BlsKeypairShare>,
+    public_id: PublicId,
+}
+
+impl NodeKeypairs {
+    /// Constructs a `NodeKeypairs` with a random Ed25519 keypair and no BLS keys.
+    pub fn new<T: CryptoRng + Rng>(rng: &mut T) -> Self {
+        let ed25519 = Ed25519Keypair::generate::<Ed25519Digest, _>(rng);
+        let name = PublicKey::Ed25519(ed25519.public).into();
+        let public_id = PublicId {
+            name,
+            ed25519: ed25519.public,
+            bls: None,
+        };
+        Self {
+            ed25519,
+            bls: None,
+            public_id,
+        }
+    }
+
+    /// Constructs a `NodeKeypairs` whose name is in the interval [start, end] (both endpoints inclusive).
+    pub fn within_range<T: CryptoRng + Rng>(start: &XorName, end: &XorName, rng: &mut T) -> Self {
+        let mut ed25519 = Ed25519Keypair::generate::<Ed25519Digest, _>(rng);
+        loop {
+            let name = PublicKey::Ed25519(ed25519.public).into();
+            if name >= *start && name <= *end {
+                let public_id = PublicId {
+                    name,
+                    ed25519: ed25519.public,
+                    bls: None,
+                };
+                return Self {
+                    ed25519,
+                    bls: None,
+                    public_id,
+                };
+            }
+            ed25519 = Ed25519Keypair::generate::<Ed25519Digest, _>(rng);
+        }
+    }
+
+    /// Returns the BLS if any, else the Ed25519.
+    pub fn public_key(&self) -> PublicKey {
+        if let Some(key) = self.public_id.bls {
+            PublicKey::BlsShare(key)
+        } else {
+            PublicKey::Ed25519(self.public_id.ed25519)
+        }
+    }
+
+    /// Returns the public keys.
+    pub fn public_id(&self) -> &PublicId {
+        &self.public_id
+    }
+
+    /// Returns the BLS public key set if any.
+    pub fn public_key_set(&self) -> Option<&PublicKeySet> {
+        self.bls.as_ref().map(|s| &s.public_key_set)
+    }
+
+    /// Signs with the BLS if any, else the Ed25519.
+    pub fn sign(&self, data: &[u8]) -> Signature {
+        if let Some(sig) = self.sign_using_bls(data) {
+            sig
+        } else {
+            self.sign_using_ed25519(data)
+        }
+    }
+
+    /// Creates a detached Ed25519 signature of `data`.
+    pub fn sign_using_ed25519<T: AsRef<[u8]>>(&self, data: T) -> Signature {
+        Signature::Ed25519(self.ed25519.sign::<Ed25519Digest>(data.as_ref()))
+    }
+
+    /// Creates a detached BLS signature share of `data` if the `self` holds a BLS keypair share.
+    pub fn sign_using_bls<T: AsRef<[u8]>>(&self, data: T) -> Option<Signature> {
+        self.bls.as_ref().map(|keys| {
+            Signature::BlsShare(SignatureShare {
+                index: keys.index,
+                share: keys.secret.inner().sign(data),
+            })
+        })
+    }
+
+    /// Sets the `NodeKeypairs`'s BLS keypair share using the provided BLS secret key share.
+    pub fn set_bls_keys(&mut self, secret_share: BlsSecretKeyShare, public_set: PublicKeySet) {
+        let public = secret_share.public_key_share();
+        let secret = SerdeSecret(secret_share);
+        self.public_id.bls = Some(public);
+        self.bls = Some(BlsKeypairShare {
+            index: 0,
+            secret,
+            public,
+            public_key_set: public_set,
+        });
+    }
+
+    /// Clears the `NodeKeypairs`'s BLS keypair share, i.e. sets it to `None`.
+    pub fn clear_bls_keys(&mut self) {
+        self.public_id.bls = None;
+        self.bls = None;
     }
 }
